+++ components/kvs_webrtc/src/kvs_webrtc.c
@@ -1076,7 +1076,7 @@ static VOID onIceCandidateHandler(UINT64 customData, PCHAR candidateJson)
                 answer_msg.message_type = WEBRTC_MESSAGE_TYPE_ANSWER;
                 STRCPY(answer_msg.peer_client_id, session->peer_id);
                 SNPRINTF(answer_msg.correlation_id, MAX_CORRELATION_ID_LEN, "%llu_%zu",
-                         GETTIME(), ATOMIC_INCREMENT(&session->correlation_id_postfix));
+                         (unsigned long long) GETTIME(), ATOMIC_INCREMENT(&session->correlation_id_postfix));
                 answer_msg.payload = payload;
                 answer_msg.payload_len = (UINT32)STRLEN(payload);
 
@@ -1452,7 +1452,7 @@ static STATUS kvs_create_and_send_offer(kvs_pc_session_t* session)
 
         // Generate correlation ID using legacy format (timestamp_counter)
         SNPRINTF(offer_msg.correlation_id, APP_WEBRTC_MAX_CORRELATION_ID_LEN, "%llu_%zu",
-                 GETTIME(), ATOMIC_INCREMENT(&session->correlation_id_postfix));
+                 (unsigned long long) GETTIME(), ATOMIC_INCREMENT(&session->correlation_id_postfix));
 
         ESP_LOGI(TAG, "Sending SDP offer for peer: %s (len=%" PRIu32 ")", session->peer_id, offer_len);
 
@@ -1907,7 +1907,7 @@ static STATUS kvs_handleOffer(kvs_pc_session_t* session, webrtc_message_t* messa
             STRCPY(answer_msg.peer_client_id, session->peer_id);
             /* Generate fresh correlation ID like legacy respondWithAnswer */
             SNPRINTF(answer_msg.correlation_id, MAX_CORRELATION_ID_LEN, "%llu_%zu",
-                     GETTIME(), ATOMIC_INCREMENT(&session->correlation_id_postfix));
+                     (unsigned long long) GETTIME(), ATOMIC_INCREMENT(&session->correlation_id_postfix));
 
             /* Serialize the SDP answer with exact-sized heap buffer */
             UINT32 answer_len = 0;
@@ -2316,7 +2316,7 @@ static VOID kvs_senderBandwidthEstimationHandler(UINT64 customData, UINT32 txByt
 
     ESP_LOGD(TAG, "TWCC adjustment for peer %s: loss=%.2f%%, video=%lluKbps, audio=%lluKbps",
              session->peer_id, session->twcc_metadata.average_packet_loss,
-             videoBitrate / 1000, audioBitrate / 1000);
+             (unsigned long long) (videoBitrate / 1000), (unsigned long long) (audioBitrate / 1000));
 }
 #endif // KVS_ENABLE_SENDER_BANDWIDTH_ESTIMATION
